[DEV 2.0] shift speech recognition + transcription logic to websocket client
Prior to this commit, the speech recognition was being handled on the server side. This creates issues when interacting with the tool,
as the user will likely utilize the microphone on their client-side device for transcription. I hooked up the Web Speech API (https://developer.mozilla.org/en-US/docs/Web/API/Web_Speech_API/Using_the_Web_Speech_API)
for the client-side speech transcription.

Note that this Web Speech API does not work on Safari. So we will have to revisit this essential feature for Safari support. Tested and works on Chrome.